% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_IMS_data.r
\name{combine_IMS_data}
\alias{combine_IMS_data}
\title{download and combine NSIDC IMS data}
\usage{
combine_IMS_data(resolution = 24, start_year = 1997, end_year = 2017,
  output_dir = tempdir())
}
\arguments{
\item{resolution}{numeric spatial resolution to use in processing (4, 24)
(default = 24)}

\item{start_year}{start year}

\item{end_year}{end year}

\item{output_dir}{where to store the final dataset (default = tempdir())}
}
\description{
download and combine NSIDC IMS data
}
\examples{

\dontrun{
# download IMS scnow data for the default range 1997 - 2017
# and store in tempdir()
combine_IMS_data()
}
}
\keyword{data,}
\keyword{ice,}
\keyword{remote}
\keyword{sensing}
\keyword{snow,}
\keyword{temporal}
