% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeference_IMS_snow_data.r
\name{georeference_IMS_snow_data}
\alias{georeference_IMS_snow_data}
\title{georeference a IMS NSIDC data file}
\usage{
georeference_IMS_snow_data(filename = "", geotiff = T)
}
\arguments{
\item{filename}{file to process}

\item{geotiff}{bolean \code{TRUE} or \code{FALSE} (default = \code{FALSE})}
}
\description{
Converted files are stored in the same directory as the initial files.
}
\examples{

\dontrun{
# download the default range 1997 - 2017
# and store in tempdir()
georeference_IMS_snow_data()
}
}
\keyword{data,}
\keyword{ice,}
\keyword{remote}
\keyword{sensing}
\keyword{snow,}
\keyword{temporal}
