% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeference_legacy_snow_statistics.r
\name{georeference_legacy_snow_statistics}
\alias{georeference_legacy_snow_statistics}
\title{download and geo-reference legacy (ASCII) NSIDC data}
\usage{
georeference_legacy_snow_statistics(output_dir = tempdir())
}
\arguments{
\item{output_dir}{where to store the final dataset (default = tempdir())}
}
\description{
download and geo-reference legacy (ASCII) NSIDC data
}
\examples{

\dontrun{
# download the data
georeference_legacy_snow_statistics()
}
}
\keyword{data,}
\keyword{ice,}
\keyword{remote}
\keyword{sensing}
\keyword{snow,}
\keyword{temporal}
